Issue #3385: extract camera-ready run-state helpers
