+++ robot_sf/benchmark/camera_ready_campaign.py
@@ -8,13 +8,10 @@ bundle for archival/release pipelines.
 from __future__ import annotations
 
 import math
-import subprocess
 import time
 from dataclasses import asdict
-from datetime import datetime
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
-from urllib.parse import urlsplit, urlunsplit
+from typing import Any
 
 import yaml
 from loguru import logger
@@ -99,6 +96,16 @@ from robot_sf.benchmark.camera_ready._route_clearance import (  # noqa: F401 - r
 from robot_sf.benchmark.camera_ready._route_clearance import (
     _build_route_clearance_warnings as _build_route_clearance_warnings_impl,
 )
+from robot_sf.benchmark.camera_ready._run_state import (  # noqa: F401 - re-exported for back-compat
+    _campaign_id,
+    _campaign_success_counters,
+    _git_context,
+    _resolve_campaign_id,
+    _resolve_execution_mode,
+    _resolve_observation_noise,
+    _resolve_path,
+    _sanitize_git_remote,
+)
 from robot_sf.benchmark.camera_ready._summaries import (  # noqa: F401 - re-exported for back-compat
     _ACTUATION_REPORT_METRICS,
     _SEED_VARIABILITY_METRICS,
@@ -143,16 +150,12 @@ from robot_sf.benchmark.fallback_policy import (
     summarize_campaign_outcome,
     summarize_campaign_status_axes,
 )
-from robot_sf.benchmark.fallback_policy import (
-    resolve_execution_mode as _resolve_benchmark_execution_mode,
-)
 from robot_sf.benchmark.latency_stress import (
     load_latency_stress_profile,
     not_available_latency_metrics,
     validate_latency_stress_profile,
 )
 from robot_sf.benchmark.observation_noise import (
-    load_observation_noise_spec,
     normalize_observation_noise_spec,
     observation_noise_hash,
 )
@@ -183,7 +186,6 @@ from robot_sf.benchmark.synthetic_actuation import (
 )
 from robot_sf.benchmark.utils import (
     _config_hash,
-    _git_hash_fallback,
     load_optional_json,
 )
 from robot_sf.common.artifact_paths import (
@@ -193,52 +195,9 @@ from robot_sf.common.artifact_paths import (
 )
 from robot_sf.nav.svg_map_parser import convert_map
 
-if TYPE_CHECKING:
-    from collections.abc import Mapping, Sequence
-
 CAMPAIGN_SCHEMA_VERSION = "benchmark-camera-ready-campaign.v1"
 DEFAULT_EPISODE_SCHEMA_PATH = Path("robot_sf/benchmark/schemas/episode.schema.v1.json")
 _normalized_kinematics_matrix = _kinematics_matrix_or_default
-
-
-def _campaign_success_counters(
-    run_entries: Sequence[Mapping[str, Any]], *, expected_core_runs: int | None = None
-) -> dict[str, Any]:
-    """Return campaign success counters, anchoring success on core planners when present."""
-    total_runs = 0
-    successful_runs = 0
-    core_total_runs = 0
-    core_successful_runs = 0
-
-    for entry in run_entries:
-        total_runs += 1
-        is_ok = str(entry.get("status", "")) == "ok"
-        if is_ok:
-            successful_runs += 1
-
-        planner_group = str((entry.get("planner") or {}).get("planner_group", "")).strip().lower()
-        if planner_group == "core":
-            core_total_runs += 1
-            if is_ok:
-                core_successful_runs += 1
-
-    if expected_core_runs is None:
-        expected_core_runs = core_total_runs
-
-    if core_total_runs:
-        success_basis = "core"
-        benchmark_success = core_successful_runs == core_total_runs == expected_core_runs
-    else:
-        success_basis = "all"
-        benchmark_success = total_runs > 0 and successful_runs == total_runs
-    return {
-        "benchmark_success": benchmark_success,
-        "benchmark_success_basis": success_basis,
-        "successful_runs": successful_runs,
-        "total_runs": total_runs,
-        "core_successful_runs": core_successful_runs,
-        "core_total_runs": core_total_runs,
-    }
 
 
 def _build_route_clearance_warnings(
@@ -266,135 +225,6 @@ def _build_route_clearance_warnings(
         )
     finally:
         _route_clearance_module.convert_map = original_convert_map
-
-
-def _resolve_execution_mode(algorithm_metadata_contract: Any) -> str:
-    """Resolve execution mode from algorithm metadata payload with legacy fallbacks.
-
-    Returns:
-        Resolved execution mode string, or ``"unknown"`` when unavailable.
-    """
-    return _resolve_benchmark_execution_mode(algorithm_metadata_contract)
-
-
-def _sanitize_git_remote(remote: str) -> str:
-    """Remove credentials from git remote URLs before persisting provenance metadata.
-
-    Returns:
-        Credential-free remote URL when parseable, otherwise original input.
-    """
-    if not remote or "://" not in remote:
-        return remote
-    try:
-        parsed = urlsplit(remote)
-    except ValueError:
-        return remote
-    if not parsed.hostname:
-        return remote
-    netloc = parsed.hostname
-    if parsed.port is not None:
-        netloc = f"{netloc}:{parsed.port}"
-    return urlunsplit((parsed.scheme, netloc, parsed.path, parsed.query, parsed.fragment))
-
-
-def _git_context() -> dict[str, str]:
-    """Collect lightweight git metadata for campaign provenance.
-
-    Returns:
-        Mapping with ``commit``, ``branch``, and ``remote`` fields.
-    """
-
-    def _run(args: list[str]) -> str:
-        """Run a git command and degrade to ``unknown`` when provenance is unavailable.
-
-        Returns:
-            str: Decoded command output, or ``"unknown"`` on command failure.
-        """
-        try:
-            out = subprocess.check_output(args, stderr=subprocess.DEVNULL)
-            return out.decode("utf-8", errors="replace").strip()
-        except (subprocess.CalledProcessError, FileNotFoundError, OSError):
-            return "unknown"
-
-    return {
-        "commit": _git_hash_fallback(),
-        "branch": _run(["git", "rev-parse", "--abbrev-ref", "HEAD"]),
-        "remote": _sanitize_git_remote(_run(["git", "config", "--get", "remote.origin.url"])),
-    }
-
-
-def _campaign_id(cfg: CampaignConfig, *, label: str | None = None) -> str:
-    """Build a unique campaign identifier from config name and wall-clock timestamp.
-
-    Returns:
-        Campaign identifier used for output directories and manifests.
-    """
-    stamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    base = _sanitize_name(cfg.name)
-    if label:
-        suffix = _sanitize_name(label)
-        return f"{base}_{suffix}_{stamp}"
-    return f"{base}_{stamp}"
-
-
-def _resolve_campaign_id(
-    cfg: CampaignConfig,
-    *,
-    label: str | None = None,
-    campaign_id: str | None = None,
-) -> str:
-    """Resolve the output campaign identifier.
-
-    Returns:
-        Explicit sanitized campaign id when provided, otherwise a timestamped id.
-    """
-    if campaign_id is not None:
-        normalized = _sanitize_name(campaign_id)
-        if not normalized:
-            raise ValueError("campaign_id must contain at least one alphanumeric character")
-        return normalized
-    return _campaign_id(cfg, label=label)
-
-
-def _resolve_path(raw_path: str | None, *, base_dir: Path) -> Path | None:
-    """Resolve optional paths relative to ``base_dir``.
-
-    Returns:
-        Absolute resolved path, or ``None`` when no path was provided.
-    """
-    if not raw_path:
-        return None
-    path = Path(raw_path)
-    if path.is_absolute():
-        return path
-
-    candidate = (base_dir / path).resolve()
-    if candidate.exists():
-        return candidate
-
-    repo_candidate = (get_repository_root() / path).resolve()
-    if repo_candidate.exists():
-        return repo_candidate
-
-    return candidate
-
-
-def _resolve_observation_noise(raw: Any, *, base_dir: Path) -> dict[str, Any] | None:
-    """Resolve an optional inline or file-backed observation-noise config.
-
-    Returns:
-        Normalized noise spec, or ``None`` when no profile is configured.
-    """
-    if raw is None:
-        return None
-    if isinstance(raw, dict):
-        return normalize_observation_noise_spec(raw)
-    if isinstance(raw, str) and raw.strip():
-        path = _resolve_path(raw, base_dir=base_dir)
-        if path is None or not path.is_file():
-            raise FileNotFoundError(f"Could not resolve observation_noise '{raw}'")
-        return load_observation_noise_spec(path)
-    raise ValueError("observation_noise must be a mapping or YAML file path")
 
 
 def _validate_campaign_config(cfg: CampaignConfig) -> None:  # noqa: C901, PLR0912, PLR0915
